requirements.txt
--extra-index-url https://pypi.anaconda.org/OpenEye/simple
OpenEye-toolkits
